requirements.txt
# LLM + Orchestration
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama
langchain-huggingface
langgraph

# Document Parsing
pymupdf==1.25.1          # PDF
python-docx==1.1.2       # DOCX
markdown==3.7            # MD

# Vector Store + Embeddings
faiss-cpu==1.9.0
sentence-transformers==3.3.1

# Observability
langsmith

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2